requirements.txt
aiohttp==3.6.3
astroid==2.4.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
common==0.1.2
discord==1.0.1
discord.py==1.5.1
discord-py-slash-command
ffmpeg
idna==2.10
imgurpython==1.1.7
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
numpy==1.19.3
opencv-python==4.4.0.46
pycparser==2.20
pylint==2.6.0
PyGithub
PyNaCl==1.3.0
requests==2.24.0
six==1.15.0
toml==0.10.2
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.11
wrapt==1.12.1
yarl==1.5.1
youtube-dl==2020.11.1.1
